Feature/armar estudio comercial

+++ app/presentacion/api/prospecto/dto/requests/registrar_prospecto_request.py
@@ -20,5 +20,5 @@ class RegistrarProspectoRequest(BaseModel):
     cantidad_subterraneos: int | None
     tiene_piscina: bool | None
     year_construccion: int | None
-    metros_cuadrados: int | None
+    metros_cuadrados: float | None
     desea_ser_contactado: bool | None
